Refactor parsers
